regmap: improve debugfs interface to dump specific addresses
The current method of cat-ing register file dumps the entire
address space. One can use dd command to dump a subrange within
the address space. However one needs to know the string length
of each line which is derived from max address, the character
length of each register entry and the format.

Provide simple means to dump a range by allowing user to specify
the start address and the count of registers. When the data is read
convert the dump address to a starting position in the file. Similarly
if the file offset goes beyond the dump range return 0 to indicate
that the data is already dumped.

Also provide means to write to a register address.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit c1fa88dd6d7b816201c98b74ad8b5a2dfc3d542e)

Change-Id: I3466ce89007d127151f6760328edad116d679db8
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>