soc: qcom: ipc_router_glink_xprt: Fix race condition during SSR
There is a race condition when a read is queued and a
subsystem restart occurs before the read can finish. The
read will call read_done on a freed handle. This fix
changes the ss_reset lock to a rwlock and extends the
critical sections in the read and write data functions.

CRs-Fixed: 969026
Signed-off-by: Chris Lew <clew@codeaurora.org>
(cherry picked from commit c1276a3118de60ee79dffe71778ab9c2cb3302aa)
Change-Id: I7726e4bbea7447ad96df725b50e4eff1ec67607f
Signed-off-by: Amandeep Singh <amansing@codeaurora.org>