drivers: soc: smd: Adapt smd to work as a rpmsg device
In smd driver we create an qcom_smd_device
for each smd_channel and pass the qcom_smd_device to the
channel clients. The qcom_smd_device and the channel clients
are tied to qcom_smd_bus. Instead of doing this, make use
of the already existing rpmsg_bus, which does exactly similar
channel enumeration logic.

Signed-off-by: Sricharan R <sricharan@codeaurora.org>