drivers: rpmsg: Make rpmsg layer generic for any channel type
Right now the rpmsg layer is tied to work with virtio based
remoteproc devices. There are cases where the remote processors's
firmware does not support virtio type channels, but the
client <-> remoteproc that rpmsg provides can be reused.

The current virtio_rpmsg_bus.c has a driver for virtio devices
and also provides rpmsg_bus for virtio device clients. This patch
seperates out the rpmsg_bus and makes it generic for any type
of channels. So any other remote proc clients can also use the
rpmsg_bus to communicate, which looks like below.

        rpmsg_channel_client -----> Any rpmsg client
                |
                |
                |
            rpmsg_bus
                |
                |
                |
        rpmsg_channel_driver  ----> Can be virtio_rpmsg_device driver
                                    or any other remoteproc_rpmsg_driver

The rpmsg_channel_driver has to add the enumerated channels to the rpmsg_bus
and provide the functions to work on the channel.

The subsequent patches converts a remoteproc driver and its
client to be based on the rpmsg_bus and deletes the local
bus implementation from the driver.

Signed-off-by: Sricharan R <sricharan@codeaurora.org>