ipq40xx: Fix SPI NAND bit flip error
In spinand_program_page function, we have come across an issue
where bit flip occurs when raw read is done on erased page.
This changes a bit from 1 to 0. When we perform wbuf[i] &= buf[j],
the bit flip is propagated since '&' operation is performed on 0
and erroneous data is written to flash in cases where actual content
is supposed to be 1. This results in ECC being calculated for the wrong
data and hence the bit flip cannot be corrected.
To avoid this case, this patch overwrites the contents of wbuf
with the intended content using memcpy.

Change-Id: I2094c336e6fb6a1dfbca6c7bf007280af696ec21
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>
(cherry picked from commit bce27994d3eaa689a50b2b4d6862b8355f12d1b1)