misc: qseecom: fix secure multiplication test crash
In qseecom driver data buffer is allocated and dma'able address
is used only for encryption & decryption test cases. once the tz
processing is done dma'able address is unmapped.  For multiplication
test, data buffer is not used and no dma mapping & unmapping
is needed.

In the existing driver, dma unmapping is done for all the tests
which results in crash. In order to fix this, check is added to
do the dma unmap only for encryption & decryption test cases.

Change-Id: I42e151a257244f1a3c9311529314a720b3cce741
Signed-off-by: Jaiganesh Narayanan <njaigane@codeaurora.org>