ipq40xx: Fix reserved memory issue
Typically, the zone size is zone_end - zone_start, which
includes the reserved regions within that zone. However,
if the reserved region is at the end of a zone, the zone
size becomes reserved_region_start - zone_start. This
results in max_high and max_low computations to go wrong,
which in turn results in a negative zone size.

Hence, prevent zone size from being assigned a negative
value in cases where end offset is lesser than the start
offset.

Change-Id: Id4bea642400173b7c487ab4eae377a6d8b72096a
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>