clk: qcom: add support for hw controlled RCG
The current driver generates stack trace during RCG update if
the RCG is off and new parent source is also disabled. For
hardware controlled RCG’s, clock is forced on during update
process and goes back to off status once switch is completed.
Since the new parent is in disabled state so update bit won’t
be cleared in this case. The check for update bit can be
skipped in this case.

Change-Id: I5e0f9b788ecce95d24bb1a6786d37597c3dc39e1
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>