watchdog: ipq: Allocate memory for PMIC dump
This patch increases crashdump allocation size
from 8K + 384K to 8K + 384K + 8K for ipq807x.

This 8K memory can be used by TZ to dump PMIC
memory. 112K memory will be left unused currently
which can be used in future based on needs.

get_order function returns the next higher order
as output. So when we pass 400K as argument, 512K
is allocated.

This patch also modifies the qcom_scm_regsave
function which allocates memory to accept a size
as parameter which it can use to allocate memory.

Till ipq806x, the requirement was to allocate 1
page for Crashdump and so it was defaulted to
PAGE_SIZE.

But from ipq807x, the requirement is not 1 page
and can vary dynamically upon needs, so we have
modified that function to accept the size as
parameter and allocate memory accordingly.

Change-Id: Id0eccf2ceca45418abed71310120711a344fce1d
Signed-off-by: speriaka <speriaka@codeaurora.org>