watchdog: ipq: Allocate ipq807x 384 KB for NSS Dump
This patch increases crashdump allocation size from
8K to 8K + 384K for ipq807x.

As SBL overwrites the NSS IMEM, TZ has to copy it to
some memory on crash before it restarts the system.

Hence reserving of 384 KB is required to copy the
NSS IMEM before restart is done.

So that, TZ can dump NSS data after the first 8K.

Change-Id: Ie8b0d228b36d7790d1a09afc9f13494ff254309e
Signed-off-by: speriaka <speriaka@codeaurora.org>