regulator: qcom: cpr: Add CPR3 NPU regulator support for IPQ807x
Add the driver for CPR3 NPU regulator for IPQ807x which will
always operate in open loop. Since, it is always going to operate
in open loop so it won’t require any register access. It will
read the initial values from fuse and use the same for different
modes. Also, it supports the MX regulator scaling with CPR3 base
framework.

Change-Id: I17d33e0286dd2749081ae6c1f8ce8b838268b1a0
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>