sound: ipq: Fixes for Pause/Resume issue
During the pause/resume case, mbox stop - mbox resume APIs are invoked.
The flag mbox_started is unset in mbox stop. Hence removed the condition
in mbox resume, where it does not resume & returns when mbox_started is not set.

Similarly the read/write pointers are deinitialized in mbox stop. This causes it
to lose track of the pointers when pause was invoked. Moved this to mbox release API.

Change-Id: I970d9e929c43abbf2659d096979f5ff6988a427e
Signed-off-by: anusha <anusharao@codeaurora.org>