i2c: qup: schedule EOT and FLUSH tags at the end of transfer
A single BAM transfer can have multiple read and write messages.
The EOT and FLUSH tags should be scheduled at the end of BAM HW
descriptors. Since the READ and WRITE can be present in any order
so for some of the cases, these tags are not being written
correctly.

Change-Id: Idacfbb5307740dbd12b35181c793672f686e27a0
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>