mtd: nand: qcom: helper function for raw read
This patch does minor code reorganization for raw reads.
Currently the raw read is required for complete page but for
subsequent patches related with erased codeword identification,
only few CW should be read. So, this patch adds helper function
and introduces the read CW bitmask which tells what are the CW
required.

Change-Id: I212a45b92d2e27587372e620a06d98561bca6f8a
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>