ipq807x: usb: gadget serial: ttyGS0 hook to dev console.
This patch registers the the usb gadget serial [ttyGS0] to the dev console.
This is needed to make sure the kernel logs are routed to the usb console.
Also the usb kernel drivers to be built as in-built kernel modules.

To avoid the re-entrance when transferring data with usb endpoint,
it introduces a kthread to do the IO transmission.
CONFIG_U_SERIAL_CONSOLE needs to be enabled for this feature to work.

CR in every LF not Implemented.

Set the bootargs to "setenv bootargs console=ttyGS0" in uboot.

Change-Id: I7e2241a0cabfff4d6c04476a1e9fb3ebbd5fe7e9
Signed-off-by: Tushar Nimkar <tnimkar@codeaurora.org>