ipq40xx: SPI NAND: Move Device ID table
Move the device ID table for SPI NAND from the common
NAND layer to the SPINAND driver. This is to avoid
cases where 2 bytes of SPI NAND ID would match with that
of a parallel NAND device. In such a case, paralled NAND
would never go into an ONFI probe since the dev ID would
match with that of SPI NAND.

Change-Id: Ib19d732a4488d41efafaed34730ebe4397fd91b6
Signed-off-by: Charanya Venkatraman <charanya@codeaurora.org>