cfg80211: Bypass channel calculation for 4.9GHz
There are some channel numbers which are same in both 5GHz and in 2.4Ghz
bands. For example, channel number 1 maps to frequency 4942MHz in 5GHz
band, but in 2.4GHz band, channel number 1 maps to frequency 2412MHz.

For a 5GHz channel number, cfg80211 channel-to-freq conversion function
may provide a 2.4GHz channel frequency. Since a 2.4GHz frequency is not
present in a 5GHz wiphy, the scan entry of 5GHz gets dropped in the
kernel.

When the wifi driver provides the scan result to Kernel it also provides
the current channel details. Therefore, bypass the channel calculation in
kernel for 5GHz band and use the channel provided by the wifi driver.

Change-Id: I6e33a425ac46c5a7171e53ec0f5de5cd7b6112f4
Signed-off-by: Abhijit Pradhan <abhijit@codeaurora.org>