ipq807x: serial: Fix msm serial driver for earlycon
Due to using --gc-section to eliminate unused sections,
earlycon table also gets eliminated. This fix will keep
the earlycon section in the kernel image.

The default console write checks whether there is a dma
operation in progress or not. But the datastructure which
holds this information is not initialized yet during earlycon
operation. Hence we are using a parameter to distinguish
between earlycon and normal console write.

Change-Id: I4d27922f8e146211780a674fcfcd8240249a339f
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>