drivers: of: Enable iommu for platform based on cmd line option
IOMMU for the devices gets enabled based on their DT configuration.
This adds a command line option to control the IOMMU enablement.
This way we can selectively enable the iommu for specific runs
only when we 'suspect' or 'debug' issues with respect to the
IOMMU enabled dma masters.

Add 'enable_iommu' in the command line to turn the feature 'ON'.

Change-Id: I269804f21037dfc13263d6f66c7aa9e9d7ede546
Signed-off-by: Sricharan R <sricharan@codeaurora.org>